refactor(frontend): data map types
Move shared data map types into `lib/data-map/types`.
