+++ frontend/src/config/assets/AssetHoverDescription.tsx
@@ -1,6 +1,6 @@
 import { FC } from 'react';
 
-import { InteractionTarget, VectorTarget } from 'state/interactions/use-interactions';
+import { InteractionTarget, VectorTarget } from 'lib/data-map/types';
 import { VectorHoverDescription } from 'map/tooltip/content/VectorHoverDescription';
 
 export const AssetHoverDescription: FC<{ hoveredObject: InteractionTarget<VectorTarget> }> = ({
