+++ frontend/src/config/assets/asset-view-layer.ts
@@ -4,10 +4,10 @@ import {
   ViewLayerDataAccessFunction,
   ViewLayerFunctionOptions,
 } from 'lib/data-map/view-layers';
+import { VectorTarget } from 'lib/data-map/types';
 import { selectableMvtLayer } from 'lib/deck/layers/selectable-mvt-layer';
 import { getAssetDataFormats } from './data-formats';
 import { ASSETS_SOURCE } from './source';
-import { VectorTarget } from 'state/interactions/use-interactions';
 
 interface ViewLayerMetadata {
   group: string;
