+++ frontend/src/details/regions/RegionDetailsContent.tsx
@@ -1,7 +1,7 @@
 import { Typography } from '@mui/material';
 import { REGIONS_METADATA } from 'config/regions/metadata';
 import { DataItem } from 'details/features/detail-components';
-import { InteractionTarget } from 'state/interactions/use-interactions';
+import { InteractionTarget } from 'lib/data-map/types';
 import { numFormat } from 'lib/helpers';
 import { FC } from 'react';
