+++ frontend/src/lib/data-map/view-layers.ts
@@ -1,7 +1,7 @@
 import { ScaleSequential } from 'd3-scale';
 import { DataLoader } from 'lib/data-loader/data-loader';
 import { Accessor } from 'lib/deck/props/getters';
-import { InteractionTarget, VectorTarget, RasterTarget } from 'state/interactions/use-interactions';
+import { InteractionTarget, VectorTarget, RasterTarget } from './types';
 
 export interface FieldSpec {
   fieldGroup: string;
